requirements.txt
boto3==1.9.230
nltk==3.4.5
pymystem3==0.2.0
PyYAML==5.1
